+++ src/lib/gtfs-to-geojson.ts
@@ -1,7 +1,5 @@
 import path from 'node:path';
 import { rm, mkdir, writeFile } from 'node:fs/promises';
-import { readFileSync } from 'node:fs';
-import { fileURLToPath } from 'node:url';
 
 import { clone, omit, uniqBy } from 'lodash-es';
 import { getRoutes, getTrips, openDb, importGtfs } from 'gtfs';
@@ -24,11 +22,7 @@ import stops from './formats/stops.js';
 import stopsBuffer from './formats/stops-buffer.js';
 import stopsDissolved from './formats/stops-dissolved.js';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const packageJson = JSON.parse(
-  readFileSync(path.join(__dirname, '../../package.json'), 'utf8'),
-);
+import packageJson from '../../package.json' with { type: 'json' };
 const { version } = packageJson;
 
 import type { Config } from '../types/global_interfaces.js';